Added text screen files for journal. Set up journal screen to dynamically change depending on entry list.

+++ app/src/main/java/com/example/myapplication/data/model/JournalEntry.kt
@@ -3,7 +3,7 @@ package com.example.myapplication.data.model
 import android.R
 
 data class JournalEntry(
-    val id: Long = 0,
+    val id: Int = 0,
     val date: String = "",
     val content: String = "",
     val complete: Boolean = false
